dbasdjangoapi/requirements.txt
django
requests
mysqlclient==2.0.1
django-rest-framework
python-dotenv
django-dotenv